requirements.txt
# ==============================================================================
# GUARDIÁN DIGITAL v2.0 - requirements.txt
# ==============================================================================

# --- Web Framework ---
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# --- HTTP Client ---
httpx==0.27.0

# --- Base de Datos ---
PyMySQL==1.1.0
cryptography==42.0.7         # Requerido por PyMySQL para SSL

# --- Google Gemini AI ---
google-generativeai==0.5.4

# --- Entorno / Config ---
python-dotenv==1.0.1

# --- Análisis de texto ---
# (Solo librería estándar: re, difflib, math)

# --- RAG / Vector Store (opcional) ---
# Instalar solo si RAG_ENABLED=true
chromadb==0.5.0
# google-generativeai ya incluye los embeddings

# --- Vision / OCR (opcional) ---
# Instalar solo si VISION_ENABLED=true
Pillow==10.3.0
pytesseract==0.3.10          # Requiere: apt install tesseract-ocr tesseract-ocr-spa

# --- Voz / Whisper (opcional) ---
# Instalar solo si VOICE_ENABLED=true
openai-whisper==20231117
ffmpeg-python==0.2.0         # Requiere: apt install ffmpeg

# --- Monitoreo del sistema ---
psutil==5.9.8
